headless-runner/requirements.txt
# Cloud services
google-cloud-storage>=2.10.0
google-auth>=2.23.0

# Data validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Utilities
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Logging
structlog>=23.2.0

# Simulator (placeholder - will be replaced with actual proprietary package)
# netgenius-simulator>=1.0.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.12.0
mypy>=1.7.0
flake8>=6.1.0
